minor code reorg, random bugfixes
