+++ src/webserver/backend/index.js
@@ -1,15 +1,16 @@
-const fs = require('fs');
 const path = require('path');
 const http = require('http');
 const https = require('https');
-const crypto = require('crypto');
 
 const express = require('express');
 const expressSession = require('express-session');
 const NedbStore = require('nedb-session-store')(expressSession);
 const SocketIo = require('socket.io');
 const bodyParser = require('body-parser');
-const pem = require('pem').promisified;
+
+const util = require('./util.js');
+const setupApi = require('./api.js');
+const Database = require('./database.sh');
 
 const soft = require('../../softconfig.js');
 
@@ -30,16 +31,21 @@ class Webserver {
   #database = undefined;
 
   // create a webserver
-  constructor(options, database, omegga) {
+  constructor(options, omegga) {
     this.port = options.port || process.env.PORT || soft.DEFAULT_PORT;
     this.options = options;
-    this.database = database;
     this.omegga = omegga;
     this.dataPath = path.join(omegga.path, soft.DATA_PATH);
 
+    // the database provides omegga with metrics, chat logs, and more
+    // to help administrators keep track of their users and server
+    this.database = new Database(options, omegga);
+
+    // https status of the server
     this.https = false;
-    this.created = this.createServer();
+    // started status of the server
     this.started = false;
+    this.created = this.createServer();
   }
 
   // create the webserver
@@ -53,8 +59,12 @@ class Webserver {
     const pickProtocol = async () => {
       if (this.options.https) {
         if (hasOpenSSL) {
-          const certFile = await this.getSSLKeys();
+          const certFile = await util.getSSLKeys(this.dataPath);
           if (certFile) {
+            // handy notification of generated ssl certs
+            if (certFile.new)
+              log('>>'.green, 'Generated new SSL certificate');
+
             this.https = true;
             return https.createServer({
               key: certFile.keys.serviceKey,
@@ -78,7 +88,7 @@ class Webserver {
 
     this.app.set('trust proxy', 1);
     const session = expressSession({
-      secret: this.getSessionSecret(),
+      secret: util.getSessionSecret(this.dataPath),
       resave: false,
       saveUninitialized: false,
       cookie: {
@@ -96,73 +106,6 @@ class Webserver {
     // setup routes and webserver
     this.initWebUI(session);
     return true;
-  }
-
-  // generate SSL certs for the https server
-  async getSSLKeys() {
-    const certsPath = path.join(this.dataPath, soft.WEB_CERTS_DATA);
-    let certData;
-    const now = Date.now();
-
-    // read cert data from json file
-    if (fs.existsSync(certsPath)) {
-      try {
-        certData = JSON.parse(fs.readFileSync(certsPath, 'utf8'));
-
-        // make sure the cert is not expired
-        if (certData.expires < now)
-          certData = undefined;
-      } catch (e) {
-        // nothing to do here - probably bad json
-      }
-    }
-
-    // otherwise generate it
-    if (!certData) {
-      // expires in half the real duration time
-      const days = 360;
-      const expires = now + (days/2) * 24 * 60 * 60 * 1000;
-      try {
-        const keys = await pem.createCertificate({ days, selfSigned: true });
-        certData = { keys, expires };
-
-        fs.writeFileSync(certsPath, JSON.stringify(certData));
-        log('>>'.green, 'Generated new SSL certificate');
-      } catch (e) {
-        // probably missing openssl or something
-      }
-    }
-
-    return certData;
-  }
-
-  // generate a session secret for the cookies
-  getSessionSecret() {
-    const tokenPath = path.join(this.dataPath, soft.WEB_SESSION_TOKEN);
-    const secretSize = 64;
-    let secret;
-    // read secret from file
-    if (fs.existsSync(tokenPath)) {
-      try {
-        secret = fs.readFileSync(tokenPath, 'utf8');
-      } catch (e) {
-        // nothing to do here - probably file is a folder or something
-      }
-    }
-
-    if (!secret) {
-      // generate a new secret
-      const buf = Buffer.alloc(secretSize);
-      secret = crypto.randomFillSync(buf).toString('hex');
-      try {
-        // write secret to file
-        fs.writeFileSync(tokenPath, secret);
-      } catch (e) {
-        // nothing to do here - maybe missing perms?
-      }
-    }
-
-    return secret;
   }
 
   // setup the web ui routes
@@ -189,69 +132,8 @@ class Webserver {
     this.app.use('/public', express.static(ASSET_PATH));
     this.app.use(bodyParser.json());
 
-    // open API is accessible without auth
-    const openApi = express.Router();
-    const api = express.Router();
-
-    // check if this is the first user in the database
-    openApi.get('/first', async (req, res) =>
-      res.json(await this.database.isFirstUser()));
-
-    // login / create admin user route
-    openApi.post('/auth', async (req, res) => {
-      // body is username and password
-      if (typeof req.body !== 'object' ||
-        typeof req.body.username !== 'string' || typeof req.body.password !== 'string') {
-        return res
-          .status(422)
-          .json({message: 'invalid body'});
-      }
-      const { username, password } = req.body;
-
-      // username regex
-      if (!username.match(/^\w{0,32}$/)) {
-        return res
-          .status(422)
-          .json({message: 'invalid body'});
-      }
-
-      // if this is the first user, create it as the admin user
-      const isFirst = await this.database.isFirstUser();
-      let user;
-      if (isFirst) {
-        user = await this.database.createAdminUser(username, username === '' ? '' : password);
-      } else {
-        user = await this.database.authUser(username, password);
-      }
-
-      if (user) {
-        req.session.userId = user._id;
-        req.session.save();
-        res.status(200).json({});
-      } else {
-        res.status(404).json({message: 'no user found'});
-      }
-    });
-
-    // kill a session
-    api.get('/logout', (req, res) => {
-      req.session.destroy(e => {
-        res.status(e ? 500 : 200).json({});
-      });
-    });
-
-    // authentication middleware for api
-    api.all(async (req, res, next) => {
-      const user = await this.database.findUserById(req.session.userId);
-      if (!user || user.isBanned)
-        return next(new Error('unauthorized'));
-      req.user = user;
-      next();
-    });
-
-    // register routes
-    this.app.use('/api/v1', openApi);
-    this.app.use('/api/v1', api);
+    // setup the api
+    setupApi(this, io);
 
     // every request goes through the index file (frontend handles 404s)
     this.app.use(async (req, res) => {
@@ -284,7 +166,6 @@ class Webserver {
     this.server.close();
     this.started = false;
   }
-
 }
 
 module.exports = Webserver;
